feat(logging): Implement better webdriver command logging (#28)
Hopefully, this covers #5. Webdriver command logs now look like this

```
20:08:14.830 |    834ms | 37f13c | NewSession
    {"browserName":"chrome"}
20:08:15.674 |      4ms | 37f13c | SetTimeouts
20:08:15.681 |    578ms | 37f13c | Go http://localhost:8081/ng1/#/interaction
20:08:16.300 |    438ms | 37f13c | FindElement
    Using css selector \'.invalid\'
    ERROR: no such element
```
or this
```
20:25:30.487 |    106ms | 73eefa | Go http://localhost:8081/ng1/#/interaction
20:25:30.620 |     25ms | 73eefa | FindElement
    Using css selector '*[id="flux"]'
    Elements: 0
20:25:30.658 |     55ms | 73eefa | ElementClick (0)
20:25:30.819 |      8ms | 73eefa | GetElementRect (0)
    {"width":88,"hCode":88,"class":"org.openqa.selenium.Dimension","height":20}
```
The second field is how long the command took. If a command has extra data associated with it, it's printed on the next line. If the response has important data (or if there's an error), it's printed on the line after that. For commands that accept an element ID, the id is put on the end of the first line in parens (`GetElementRect (0)` for example).

Commands are logged when the response from selenium is received, though the timestamp reflects when the proxy saw the command from upstream.

For commands that the proxy doesn't understand, it just shows the URL of the command and doesn't print any of the extra data.