feat(highlight): Add highlight delay.
When starting the proxy, you can now pass a --highlightDelay option that
specifies a number of seconds to wait before forwarding a page modifying
command, such as clicking an element or sending text. The element the
action is about to be performed on will be highlihted during the delay
period.